Improve cancellation dialog buttons
This change makes the button labels more explicit (to delete, you have to click a button that says "Delete"), and makes the cancel button a different color, so it is easier to differentiate the buttons and their effects.

+++ src/Components/ProfileListPage.js
@@ -74,6 +74,9 @@ export default function ProfileListPage() {
       content: "Deleting a watchlist is permanent. There is no undo.",
       titleProps: { sx: { fontFamily: "interExtraBold" } },
       contentProps: { sx: { fontFamily: "interMedium" } },
+      confirmationText: "Delete",
+      cancellationButtonProps: { color: "inherit" },
+      cancellationText: "Cancel",
     }).then(() => {
       deleteFn();
       navigate(`/profile/${userId}`);
